adding users.html page

+++ Controllers/TasksController.cs
@@ -1,6 +1,8 @@
 using Microsoft.AspNetCore.Mvc;
 using myTasks.Interfaces;
 using Task = myTasks.Models.Task;
+using Microsoft.AspNetCore.Authorization;
+
 
 namespace myTasks.Controllers;
 
@@ -9,37 +11,50 @@ namespace myTasks.Controllers;
 public class TasksController : ControllerBase
 {
     ITasksService TasksService;
+
     public TasksController(ITasksService TasksService)
     {
         this.TasksService = TasksService;
     }
+
     [HttpGet]
+    [Authorize(Policy = "USER")]
     public ActionResult<List<Task>> Get()
     {
-        return TasksService.GetAll();
+        return TasksService.GetAll(int.Parse(User.FindFirst("id").Value));
     }
 
     [HttpGet("{id}")]
+    [Authorize(Policy = "USER")]
     public ActionResult<Task> Get(int id)
     {
         var Task = TasksService.GetById(id);
         if (Task == null)
             return NotFound();
+        if(!CheckAthorization(id))
+            return Unauthorized();      
         return Task;
     }
 
     [HttpPost]
+    [Authorize(Policy = "USER")]
     public ActionResult Post(Task newTask)
     {
+        newTask.OwnerId=int.Parse(User.FindFirst("id").Value);
         var newId = TasksService.Add(newTask);
-
         return CreatedAtAction("Post", 
             new {id = newId}, TasksService.GetById(newId));
     }
 
     [HttpPut("{id}")]
+    [Authorize(Policy = "USER")]
     public ActionResult Put(int id,Task newTask)
     {
+        newTask.OwnerId=int.Parse(User.FindFirst("id").Value);
+        if(TasksService.GetById(id)==null)
+            return NotFound();
+        if(!CheckAthorization(id))
+            return Unauthorized();
         var result = TasksService.Update(id, newTask);
         if (!result)
         {
@@ -47,14 +62,20 @@ public class TasksController : ControllerBase
         }
         return NoContent();
     }
+
     [HttpDelete("{id}")]
-        public ActionResult Delete(int id)
+    [Authorize(Policy = "USER")]
+    public ActionResult Delete(int id)
     {
-        var result = TasksService.Delete(id);
-        if (!result)
-        {
-            return BadRequest();
-        }
+        if(TasksService.GetById(id)==null)
+            return NotFound();
+        if(!CheckAthorization(id))
+            return Unauthorized();
+        TasksService.Delete(id);
         return NoContent();
     } 
+
+    private  bool CheckAthorization(int taskId) {
+        return TasksService.GetById(taskId).OwnerId==int.Parse(User.FindFirst("id").Value);
+    }   
 }
